Computer Vision
Project description
The supermarket chain Good Seed would like to explore whether Data Science can help them adhere to alcohol laws by making sure they do not sell alcohol to people underage. You are asked to conduct that evaluation, so as you set to work, keep the following in mind:
The shops are equipped with cameras in the checkout area which are triggered when a person is buying alcohol
Computer vision methods can be used to determine age of a person from a photo
The task then is to build and evaluate a model for verifying people's age
To start working on the task, you'll have a set of photographs of people with their ages indicated.
Project Instructions
Pass a quiz to verify your understanding of the project statement.
Perform exploratory data analysis to get an overall impression of the dataset.
Train and evaluate the model (it needs to be done on the GPU platform).
Combine your code, output and findings (from the previous points) in the final Jupyter notebook.
Make conclusions of the model evaluation, add them to the notebook.
Project reviewers will review your final notebook.